requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
CacheControl==0.14.0
cachetools==5.3.2
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.2
daphne==4.0.0
distlib==0.3.7
Django==4.2.1
django-cors-headers==4.0.0
django-filter==21.1
django-prometheus==2.3.1
django-redis==5.4.0
django-rest-swagger==2.2.0
django-smtp-ssl==1.0
django-ssl-smtp==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.2.0
filelock==3.13.1
firebase-admin==6.4.0
google-api-core==2.17.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
jsonfield==3.1.0
kombu==5.4.0
MarkupSafe==2.1.3
MouseInfo==0.1.3
msgpack==1.0.7
mss==9.0.1
numpy==2.0.1
openapi-codec==1.3.2
packaging==23.2
pandas==2.2.2
pbr==5.11.1
pika==1.3.2
pillow==10.3.0
platformdirs==3.11.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
PyGetWindow==0.0.9
PyJWT==2.7.0
PyMsgBox==1.0.9
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.9.0.post0
python-decouple==3.8
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
razorpay==1.4.2
redis==5.0.3
requests==2.31.0
rsa==4.9
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
tomli==2.0.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.13
zope.interface==6.1